+++ blockingpy/mlpack_blocker.py
@@ -4,9 +4,14 @@ import logging
 from typing import Any
 
 import pandas as pd
-from mlpack import knn, lsh
+
+try:
+    from mlpack import knn, lsh
+except ImportError:
+    knn = lsh = None
 
 from .base import BlockingMethod
+from .data_handler import DataHandler
 from .helper_functions import rearrange_array
 
 logger = logging.getLogger(__name__)
@@ -54,8 +59,8 @@ class MLPackBlocker(BlockingMethod):
 
     def block(
         self,
-        x: pd.DataFrame,
-        y: pd.DataFrame,
+        x: DataHandler,
+        y: DataHandler,
         k: int,
         verbose: bool | None,
         controls: dict[str, Any],
@@ -115,6 +120,10 @@ class MLPackBlocker(BlockingMethod):
         """
         logger.setLevel(logging.INFO if verbose else logging.WARNING)
 
+        self.x_columns = list(x.cols)
+        X = x.to_dense()
+        Y = y.to_dense()
+
         self.algo = controls.get("algo", "lsh")
         self._check_algo(self.algo)
         seed = controls.get("random_seed", None)
@@ -123,12 +132,12 @@ class MLPackBlocker(BlockingMethod):
         else:
             k_search = controls["kd"].get("k_search")
 
-        if k_search > x.shape[0]:
+        if k_search > X.shape[0]:
             original_k_search = k_search
-            k_search = min(k_search, x.shape[0])
+            k_search = min(k_search, X.shape[0])
             logger.warning(
                 f"k_search ({original_k_search}) is larger than the number of reference points "
-                f"({x.shape[0]}). Adjusted k_search to {k_search}."
+                f"({X.shape[0]}). Adjusted k_search to {k_search}."
             )
 
         logger.info(f"Initializing MLPack {self.algo.upper()} index...")
@@ -136,8 +145,8 @@ class MLPackBlocker(BlockingMethod):
         if self.algo == "lsh":
             query_result = lsh(
                 k=k_search,
-                query=y,
-                reference=x,
+                query=Y,
+                reference=X,
                 verbose=verbose,
                 seed=seed,
                 bucket_size=controls["lsh"].get("bucket_size"),
@@ -149,8 +158,8 @@ class MLPackBlocker(BlockingMethod):
         else:
             query_result = knn(
                 k=k_search,
-                query=y,
-                reference=x,
+                query=Y,
+                reference=X,
                 verbose=verbose,
                 seed=seed,
                 algorithm=controls["kd"].get("algorithm"),
@@ -172,7 +181,7 @@ class MLPackBlocker(BlockingMethod):
 
         result = pd.DataFrame(
             {
-                "y": range(y.shape[0]),
+                "y": range(Y.shape[0]),
                 "x": indices[:, k - 1],
                 "dist": distances[:, k - 1],
             }